Move `std::io::util` to `core::io`
